Revert "Change parseInt to parseFloat and remove unnecessary function"
This reverts commit cb4b7c37dd5b1af3b2fb56d4c9a346fec89caebe.

+++ client/calculate.js
@@ -12,9 +12,9 @@ function parseFormula(formula) {
       currentNumber.push(element)
     } else if (index === array.length - 1) {
       currentNumber.push(element)
-      parsedFormula.push(parseFloat(currentNumber.join('')))
+      parsedFormula.push(parseInt(currentNumber.join('')))
     } else {
-      parsedFormula.push(parseFloat(currentNumber.join('')))
+      parsedFormula.push(parseInt(currentNumber.join('')))
       currentNumber.length = 0
       parsedFormula.push(element)
     }
